pyproject.toml
[project]
name = "claimlightning"
version = "0.1.0"
description = ""
authors = [
    {name = "github.com/Stosan"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "openai (>=1.100.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pymongo (>=4.14.0,<5.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiofiles (>=24.1.0,<25.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
